[#474] Add type safe BlockHeight API
 - Replace int with type safe BlockHeight(Long) object
 - Fix down casting bug, as BlockHeight is uint32 but Java only supports int32 or int64
 - Rename WalletBirthday to Checkpoint and hide from the public API

+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/db/CompactBlockDb.kt
@@ -43,11 +43,11 @@ interface CompactBlockDao {
     suspend fun insert(block: List<CompactBlockEntity>)
 
     @Query("DELETE FROM compactblocks WHERE height > :height")
-    suspend fun rewindTo(height: Int)
+    suspend fun rewindTo(height: Long)
 
     @Query("SELECT MAX(height) FROM compactblocks")
-    suspend fun latestBlockHeight(): Int
+    suspend fun latestBlockHeight(): Long
 
     @Query("SELECT data FROM compactblocks WHERE height = :height")
-    suspend fun findCompactBlock(height: Int): ByteArray?
+    suspend fun findCompactBlock(height: Long): ByteArray?
 }
